SiNoS update: version 1.0.0 Please, note that this version is not compatible with the previous version (0.4.4).
Release notes:

- The situation service is now able to manage channels' closure, as well
as to disable publishers' ability of creating new channels. The service
is not yet able to create new channels by itself.

- Situation service users are now able to list the existing channels,
publishers and subscribers. After initialising the situation service,
type "help" for more information.

- Publishers are not able to close channels anymore. Channels' closure
is only possible through service's command line.

- Both publishers and subscribers are now able to retrieve a list of
existing channels in the connected service. It is possible by means of
the 'getChannels' method in the SituationChannel class (for publishers)
and ServiceConnection class (for subscribers)

- As previously mentioned, the situation service can now refuse create
new channels. In order to check if it is or not allowed to create new
channels, publishers can invoke the 'canCreateChannel' method in the
SituationChannel class.

- In the previous version, subscribers were disconnected from channels
that had no publishers registered. This bug has been fixed.


Please, feel free to report bugs and suggest improvements.